+++ api/lib/source/cosmosV2-source.js
@@ -138,14 +138,10 @@ class CosmosV2API extends CosmosV0API {
       [].concat(...results)
     )
 
-    return this.reducers.transactionsReducerV2(
-      this.networkId,
-      txs,
-      this.reducers
-    )
+    return this.reducers.transactionsReducerV2(this.network, txs, this.reducers)
   }
 
-  async getRewards(delegatorAddress, fiatCurrency) {
+  async getRewards(delegatorAddress, fiatCurrency, network) {
     this.checkAddress(delegatorAddress)
     const result = await this.query(
       `distribution/delegators/${delegatorAddress}/rewards`
@@ -158,7 +154,8 @@ class CosmosV2API extends CosmosV0API {
       this.store.validators,
       fiatCurrency,
       this.calculateFiatValue && this.calculateFiatValue.bind(this),
-      this.reducers
+      this.reducers,
+      network
     )
   }
 }