+++ app/src/vuex/modules/wallet.js
@@ -1,13 +1,13 @@
 // creates a cosmos addres for the network desired
-function getCosmosAddressCreator(bech32Prefix) {
+function getCosmosAddressCreator(bech32Prefix, HDPath, curve) {
   return async (seedPhrase) => {
     const { getNewWalletFromSeed } = await import("@lunie/cosmos-keys")
-    return getNewWalletFromSeed(seedPhrase, bech32Prefix)
+    return getNewWalletFromSeed(seedPhrase, bech32Prefix, HDPath, curve)
   }
 }
 
 // creates a polkadot address
-async function createPolkadotAddress(seedPhrase, addressPrefix) {
+async function createPolkadotAddress(seedPhrase, network, curve) {
   const [{ Keyring }] = await Promise.all([
     import("@polkadot/api"),
     import("@polkadot/wasm-crypto").then(async ({ waitReady }) => {
@@ -20,8 +20,8 @@ async function createPolkadotAddress(seedPhrase, addressPrefix) {
   ])
 
   const keyring = new Keyring({
-    ss58Format: Number(addressPrefix),
-    type: "sr25519",
+    ss58Format: Number(network.address_prefix),
+    type: curve,
   })
   const newPair = keyring.addFromUri(seedPhrase)
 
@@ -29,19 +29,22 @@ async function createPolkadotAddress(seedPhrase, addressPrefix) {
     cosmosAddress: newPair.address,
     publicKey: newPair.publicKey,
     seedPhrase,
+    curve,
   }
 }
 
-export async function getWallet(seedPhrase, network) {
+export async function getWallet(seedPhrase, network, HDPath, curve) {
   switch (network.network_type) {
     case "cosmos": {
       const addressCreator = await getCosmosAddressCreator(
-        network.address_prefix
+        network.address_prefix,
+        HDPath,
+        curve
       )
       return await addressCreator(seedPhrase)
     }
     case "polkadot": {
-      return await createPolkadotAddress(seedPhrase, network.address_prefix)
+      return await createPolkadotAddress(seedPhrase, network, curve)
     }
     default:
       throw new Error(