+++ app/tests/unit/specs/components/common/AppMenu.spec.js
@@ -54,6 +54,7 @@ describe(`AppMenu`, () => {
   it(`call dispatch to sign the user out`, () => {
     const $store = { dispatch: jest.fn() }
     const self = {
+      address: `cosmos1`,
       network: `la-red-feliz`,
       $store,
       $router: { push: jest.fn() },